New LongListSelector
Added SelectedItem dependency property to subclass of LongListSelector
for easier binding.
Sets it to 'null' when navigating away so that the user can select same
item again. It did nothing before.
Added a trigger to fire when new items should be loaded for infinite
scrolling but it's not doing the actual loading yet.